+++ src/test/java/io/kontur/layers/controller/CollectionsSearchPostIT.java
@@ -241,6 +241,34 @@ public class CollectionsSearchPostIT extends AbstractIntegrationTest {
 
     @Test
     @WithMockUser("owner_3")
+    public void testGetOwnedAndPublicCollections_2() throws Exception {
+        //GIVEN
+        testDataMapper.insertLayer(buildLayerN(1));
+        Layer layer2 = buildLayerN(2);
+        ReflectionTestUtils.setField(layer2, "isPublic", false);
+        testDataMapper.insertLayer(layer2);
+        Layer layer3 = buildLayerN(3);
+        ReflectionTestUtils.setField(layer3, "isPublic", false);
+        testDataMapper.insertLayer(layer3);
+        //WHEN
+        String response = mockMvc.perform(post("/collections/search")
+                        .contentType(APPLICATION_JSON)
+                        .content("{\"collectionOwner\": \"ANY\"}"))
+                .andDo(print())
+                .andExpect(status().isOk())
+                .andExpect(content().contentType(APPLICATION_JSON))
+                .andReturn().getResponse().getContentAsString();
+
+        //THEN
+        final DocumentContext json = JsonPath.parse(response);
+        assertThat(json.read("$.collections"), hasSize(2));
+        assertThat(json.read("$.numberMatched"), is(2));
+        assertThat(json.read("$.numberReturned"), is(2));
+        assertThat(json.read("$.collections[*].id"), containsInAnyOrder("pubId_1", "pubId_3"));
+    }
+
+    @Test
+    @WithMockUser("owner_3")
     public void testGetOnlyOwnedCollection() throws Exception {
         //GIVEN
         testDataMapper.insertLayer(buildLayerN(1));
@@ -253,7 +281,7 @@ public class CollectionsSearchPostIT extends AbstractIntegrationTest {
         //WHEN
         String response = mockMvc.perform(post("/collections/search")
                         .contentType(APPLICATION_JSON)
-                        .content("{\"ownedByUser\": true}"))
+                        .content("{\"collectionOwner\": \"ME\"}"))
                 .andDo(print())
                 .andExpect(status().isOk())
                 .andExpect(content().contentType(APPLICATION_JSON))
@@ -265,6 +293,34 @@ public class CollectionsSearchPostIT extends AbstractIntegrationTest {
         assertThat(json.read("$.numberMatched"), is(1));
         assertThat(json.read("$.numberReturned"), is(1));
         assertThat(json.read("$.collections[*].id"), containsInAnyOrder("pubId_3"));
+    }
+
+    @Test
+    @WithMockUser("owner_3")
+    public void testGetOthersCollections() throws Exception {
+        //GIVEN
+        testDataMapper.insertLayer(buildLayerN(1));
+        Layer layer2 = buildLayerN(2);
+        ReflectionTestUtils.setField(layer2, "isPublic", false);
+        testDataMapper.insertLayer(layer2);
+        Layer layer3 = buildLayerN(3);
+        ReflectionTestUtils.setField(layer3, "isPublic", false);
+        testDataMapper.insertLayer(layer3);
+        //WHEN
+        String response = mockMvc.perform(post("/collections/search")
+                        .contentType(APPLICATION_JSON)
+                        .content("{\"collectionOwner\": \"NOT_ME\"}"))
+                .andDo(print())
+                .andExpect(status().isOk())
+                .andExpect(content().contentType(APPLICATION_JSON))
+                .andReturn().getResponse().getContentAsString();
+
+        //THEN
+        final DocumentContext json = JsonPath.parse(response);
+        assertThat(json.read("$.collections"), hasSize(1));
+        assertThat(json.read("$.numberMatched"), is(1));
+        assertThat(json.read("$.numberReturned"), is(1));
+        assertThat(json.read("$.collections[*].id"), containsInAnyOrder("pubId_1"));
     }
 
     @Test